Added extra checks and error messages

+++ build.sh
@@ -8,23 +8,35 @@ if [ $USER = "root" ]; then
 	echo ''
 
 	if [ -d /usr/lib/gwallpaper ]; then
-		echo '==> Removing old library folder'
+		echo '==> Removing old library folder...'
 		rm -r /usr/lib/gwallpaper && echo '  -> Old library folder removed!' || echo '  -> Old library folder not removed :('
 	fi
 
-	echo '==> Making library directory'
-	mkdir -p /usr/lib/gwallpaper && echo '  -> Directory created!' || echo '  -> Directory not created :('
+	if [[ ! -d /use/lib/gwallpaper ]]; then
+		echo '==> Creating library directory...'
+		mkdir -p /usr/lib/gwallpaper && echo '  -> Directory created!' || echo '  -> Directory not created :('
+	else
+		echo "ERROR: library directory already exists!" && exit
+	fi
 
-	echo '==> Copying source folder into /usr/lib/gwallpaper'
-	cp -r ./ /usr/lib/gwallpaper && echo '  -> Source folder copied!' || echo '  -> Source folder not copied :('
+	if [[ -d /usr/lib/gwallpaper ]]; then
+		echo '==> Copying source folder into /usr/lib/gwallpaper...'
+		cp -r ./ /usr/lib/gwallpaper && echo '  -> Source folder copied!' || echo '  -> Source folder not copied :('
+	else
+		echo "ERROR: unable to copy source folder into /usr/lib/gwallpaper, directory does not exist!" && exit
+	fi
 
 	if [ -e /usr/bin/wallpaper ]; then
-		echo '==> Removing old link'
+		echo '==> Removing old link...'
 		rm /usr/bin/wallpaper && echo '  -> Old link removed!' || echo '  -> Old link not removed :('
 	fi
 
-	echo '==> Linking script into /usr/bin'
-	ln -s /usr/lib/gwallpaper/gwallpaper.sh /usr/bin/wallpaper && chmod +x /usr/bin/wallpaper && echo '  -> Linking complete!' || echo '  -> Linking failed :('
+	if [[ ! -e /usr/bin/wallpaper ]]; then
+		echo '==> Linking script into /usr/bin...'
+		ln -s /usr/lib/gwallpaper/gwallpaper.sh /usr/bin/wallpaper && chmod +x /usr/bin/wallpaper && echo '  -> Linking complete!' || echo '  -> Linking failed :('
+	else
+		echo "ERROR: unable to link script into /usr/bin, file already exists!" && exit
+	fi
 
 	echo ''
 	echo '== Install complete! =='
